api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
httpx==0.27.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
vercel-blob==0.0.8
